Rename `createAndSaveNotifications` to `createNotifications`
